+++ integration-tests/api/__tests__/admin/product.js
@@ -41,6 +41,7 @@ describe("/admin/products", () => {
       const manager = dbConnection.manager;
       await manager.query(`DELETE FROM "product_option_value"`);
       await manager.query(`DELETE FROM "product_option"`);
+      await manager.query(`DELETE FROM "image"`);
       await manager.query(`DELETE FROM "money_amount"`);
       await manager.query(`DELETE FROM "product_variant"`);
       await manager.query(`DELETE FROM "product"`);
@@ -61,6 +62,7 @@ describe("/admin/products", () => {
         title: "Test product",
         description: "test-product-description",
         type: { value: "test-type" },
+        images: ["test-image.png", "test-image-2.png"],
         collection_id: "test-collection",
         tags: [{ value: "123" }, { value: "456" }],
         options: [{ title: "size" }, { title: "color" }],
@@ -90,6 +92,15 @@ describe("/admin/products", () => {
         expect.objectContaining({
           title: "Test product",
           handle: "test-product",
+          images: expect.arrayContaining([
+            expect.objectContaining({
+              url: "test-image.png",
+            }),
+            expect.objectContaining({
+              url: "test-image-2.png",
+            }),
+          ]),
+          thumbnail: "test-image.png",
           tags: [
             expect.objectContaining({
               value: "123",
@@ -136,13 +147,15 @@ describe("/admin/products", () => {
       );
     });
 
-    it("updates a product (update tags, delete collection, delete type)", async () => {
+    it("updates a product (update tags, delete collection, delete type, replaces images)", async () => {
       const api = useApi();
 
       const payload = {
         collection_id: null,
         type: null,
         tags: [{ value: "123" }],
+        images: ["test-image-2.png"],
+        type: { value: "test-type-2" },
       };
 
       const response = await api
@@ -159,6 +172,12 @@ describe("/admin/products", () => {
 
       expect(response.data.product).toEqual(
         expect.objectContaining({
+          images: expect.arrayContaining([
+            expect.objectContaining({
+              url: "test-image-2.png",
+            }),
+          ]),
+          thumbnail: "test-image-2.png",
           tags: [
             expect.objectContaining({
               value: "123",
@@ -166,6 +185,9 @@ describe("/admin/products", () => {
           ],
           type: null,
           collection: null,
+          type: expect.objectContaining({
+            value: "test-type-2",
+          }),
         })
       );
     });