+++ integration-tests/api/helpers/product-seeder.js
@@ -52,7 +52,7 @@ module.exports = async (connection, data = {}) => {
     tax_rate: 0,
   });
 
-  await manager.insert(Product, {
+  const p = manager.create(Product, {
     id: "test-product",
     title: "Test product",
     profile_id: defaultProfile.id,
@@ -81,6 +81,10 @@ module.exports = async (connection, data = {}) => {
     ],
     options: [{ id: "test-option", title: "Default value" }],
   });
+
+  p.images = [image];
+
+  await manager.save(p);
 
   await manager.insert(ProductVariant, {
     id: "test-variant",