+++ packages/medusa/src/api/routes/store/carts/update-cart.js
@@ -3,6 +3,69 @@ import { Validator, MedusaError } from "medusa-core-utils"
 
 import { defaultFields, defaultRelations } from "./"
 
+/**
+ * @oas [post] /store/carts/{id}
+ * operationId: PostCartsCart
+ * summary: Update a Cart"
+ * description: "Updates a Cart."
+ * parameters:
+ *   - (path) id=* {string} The id of the Cart.
+ * requestBody:
+ *   content:
+ *     application/json:
+ *       schema:
+ *         properties:
+ *           region_id:
+ *             type: string
+ *             description: The id of the Region to create the Cart in.
+ *           country_code:
+ *             type: string
+ *             description: "The 2 character ISO country code to create the Cart in."
+ *           email:
+ *             type: string
+ *             description: "An email to be used on the Cart."
+ *           billing_address:
+ *             description: "The Address to be used for billing purposes."
+ *             anyOf:
+ *               - $ref: "#/components/schemas/address"
+ *           shipping_address:
+ *             description: "The Address to be used for shipping."
+ *             anyOf:
+ *               - $ref: "#/components/schemas/address"
+ *           gift_cards:
+ *             description: "An array of Gift Card codes to add to the Cart."
+ *             type: array
+ *             items:
+ *               properties:
+ *                 code:
+ *                   description: "The code that a Gift Card is identified by."
+ *                   type: string
+ *           discounts:
+ *             description: "An array of Discount codes to add to the Cart."
+ *             type: array
+ *             items:
+ *               properties:
+ *                 code:
+ *                   description: "The code that a Discount is identifed by."
+ *                   type: string
+ *           customer_id:
+ *             description: "The id of the Customer to associate the Cart with."
+ *             type: string
+ *           context:
+ *             description: "An optional object to provide context to the Cart."
+ *             type: object
+ * tags:
+ *   - Cart
+ * responses:
+ *   200:
+ *     description: OK
+ *     content:
+ *       application/json:
+ *         schema:
+ *           properties:
+ *             cart:
+ *               $ref: "#/components/schemas/cart"
+ */
 export default async (req, res) => {
   const { id } = req.params
 
@@ -25,6 +88,7 @@ export default async (req, res) => {
       })
       .optional(),
     customer_id: Validator.string().optional(),
+    context: Validator.object().optional(),
   })
 
   const { value, error } = schema.validate(req.body)