+++ packages/medusa/src/api/routes/store/index.js
@@ -10,8 +10,10 @@ import orderRoutes from "./orders"
 import customerRoutes from "./customers"
 import shippingOptionRoutes from "./shipping-options"
 import regionRoutes from "./regions"
+import returnRoutes from "./returns"
 import swapRoutes from "./swaps"
 import variantRoutes from "./variants"
+import giftCardRoutes from "./gift-cards"
 
 const route = Router()
 
@@ -37,6 +39,8 @@ export default (app, container, config) => {
   regionRoutes(route)
   swapRoutes(route)
   variantRoutes(route)
+  returnRoutes(route)
+  giftCardRoutes(route)
 
   return app
 }