+++ packages/medusa/src/api/routes/store/orders/get-order-by-cart.js
@@ -1,5 +1,24 @@
 import { defaultFields, defaultRelations } from "."
 
+/**
+ * @oas [get] /orders/cart/{cart_id}
+ * operationId: GetOrdersOrderCartId
+ * summary: Retrieves Order by Cart id
+ * description: "Retrieves an Order by the id of the Cart that was used to create the Order."
+ * parameters:
+ *   - (path) cart_id=* {string} The id of Cart.
+ * tags:
+ *   - Order
+ * responses:
+ *   200:
+ *     description: OK
+ *     content:
+ *       application/json:
+ *         schema:
+ *           properties:
+ *             order:
+ *               $ref: "#/components/schemas/order"
+ */
 export default async (req, res) => {
   const { cart_id } = req.params