+++ packages/medusa/src/api/routes/store/shipping-options/list-shipping-options.js
@@ -1,5 +1,26 @@
 import { Validator, MedusaError } from "medusa-core-utils"
 
+/**
+ * @oas [get] /shipping-options/{cart_id}
+ * operationId: GetShippingOptionsCartId
+ * summary: Retrieve Shipping Options for Cart
+ * description: "Retrieves a list of Shipping Options available to a cart."
+ * parameters:
+ *   - (path) cart_id {string} The id of the Cart.
+ * tags:
+ *   - Shipping Option
+ * responses:
+ *   200:
+ *     description: OK
+ *     content:
+ *       application/json:
+ *         schema:
+ *           properties:
+ *             shipping_options:
+ *               type: array
+ *               items:
+ *                 $ref: "#/components/schemas/shipping_option"
+ */
 export default async (req, res) => {
   const schema = Validator.object().keys({
     cart_id: Validator.string().required(),